requirements.txt
flask~=3.0.0
neomodel~=5.2.0
pytest~=7.4.3
pymongo~=4.6.1
argon2-cffi
flask-login
